requirements.txt
streamlit
streamlit_authenticator
streamlit-aggrid
# streamlit.components.v1
plotly.express
openpyxl
xlsxwriter
pandas
numpy
requests
beautifulsoup4